Separate out utility crap
